pyproject.toml
[tool.poetry]
name = "swg-escolinha-futebol"
version = "0.1.0"
description = ""
authors = ["EuricoDNJR <massaultyplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
firebase-admin = "^6.5.0"
fastapi = "^0.110.3"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
peewee = "^3.17.3"
pytz = "^2024.1"
pycryptodome = "^3.20.0"
pyrebase4 = "^4.7.1"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.130.0"
apscheduler = "^3.10.4"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"

[tool.poetry.dev-dependencies]
faker = "^18.10.1"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
